The review pass, applied and priced: framing pairs, properties, enums, 246 labels - all free
Five changes from the first human read of the document-scale emitted shape.
PushScope(tag) makes every message field accept length or group framing through
paired labels over one body (an end-group sentinel is the start tag plus one, so
the group arm is free to derive), with the run loop comparing a last local
rather than a constant - and the StartSubItem veneer collapses onto it. The DTOs
now expose auto-properties (a field-assigning benchmark measures a capability
the emitted code will not have) and real enums; every case label says what it is,
unconditionally, because comments are free in every configuration. Scalar
wire-type tolerance went in as a measured A/B: 86 added labels, 160 to 246, and
the sparse-switch concern did not materialize - the whole pass nets -0.02us on
net10 and +0.21us on net472, so tolerant is simply the default and the
strict-mode knob stays a design with no trigger. Also recorded: packed/unpacked
and length/group are spec, not tolerance; hierarchies defer construction (the
inheritance-brick note); and the protogen hasBits idea, parked where it belongs.
Both halves of the milestone gate are now done.

+++ src/NanoBench/DescriptorParseBenchmarks.cs
@@ -143,7 +143,7 @@ public class DescriptorParseBenchmarks
                     {
                         c.OptionsObjects++;
                         if (m.Options.Deprecated == true) c.BoolTrues++;
-                        c.ScalarSum += m.Options.IdempotencyLevel ?? 0;
+                        c.ScalarSum += (int)(m.Options.IdempotencyLevel ?? 0);
                         c.Uninterpreted += m.Options.UninterpretedOptions.Count;
                     }
                 }
@@ -156,7 +156,7 @@ public class DescriptorParseBenchmarks
                 Census.Chars(c, o.CsharpNamespace); Census.Chars(c, o.SwiftPrefix);
                 Census.Chars(c, o.PhpClassPrefix); Census.Chars(c, o.PhpNamespace);
                 Census.Chars(c, o.PhpMetadataNamespace); Census.Chars(c, o.RubyPackage);
-                if (o.OptimizeFor is { } om) c.ScalarSum += om;
+                if (o.OptimizeFor is { } om) c.ScalarSum += (int)om;
                 foreach (var b in new[] { o.JavaMultipleFiles, o.JavaGenerateEqualsAndHash,
                     o.JavaStringCheckUtf8, o.CcGenericServices, o.JavaGenericServices,
                     o.PyGenericServices, o.PhpGenericServices, o.Deprecated, o.CcEnableArenas })
@@ -211,14 +211,14 @@ public class DescriptorParseBenchmarks
             c.FieldNumberSum += fd.Number ?? 0;
             Census.Chars(c, fd.Name); Census.Chars(c, fd.TypeName); Census.Chars(c, fd.Extendee);
             Census.Chars(c, fd.DefaultValue); Census.Chars(c, fd.JsonName);
-            if (fd.Label is { } l) c.ScalarSum += l;
-            if (fd.Type is { } t) c.ScalarSum += t;
+            if (fd.Label is { } l) c.ScalarSum += (int)l;
+            if (fd.Type is { } t) c.ScalarSum += (int)t;
             c.ScalarSum += fd.OneofIndex ?? 0;
             if (fd.Proto3Optional == true) c.BoolTrues++;
             if (fd.Options is { } o)
             {
                 c.OptionsObjects++;
-                c.ScalarSum += (o.Ctype ?? 0) + (o.Jstype ?? 0);
+                c.ScalarSum += (int)(o.Ctype ?? 0) + (int)(o.Jstype ?? 0);
                 foreach (var b in new[] { o.Packed, o.Lazy, o.Deprecated, o.Weak })
                 {
                     if (b == true) c.BoolTrues++;
